requirements.txt
streamlit
tensorflow==2.19.0
ultralytics
opencv-python-headless
numpy
pillow